MDL-66979: Switch to php-webdriver/webdriver
To facilitate this we use the oleg-andreyev/MinkPhpWebDriver Mink Driver
rather than any provided by the minkphp project, which have yet to add
support for php-webdrier.

The oleg-andreyev driver is largely complete, but there are a handful of
bugs waiting to be merged. As a result, this branch currently follows my
fork of the driver which is a branch with all of the pull requests
applied. These pull requests are:
- https://github.com/oleg-andreyev/MinkPhpWebDriver/pull/8
  Opening, closing, and re-opening a window with the same name attempts
  to reopen the closed window.
- https://github.com/oleg-andreyev/MinkPhpWebDriver/pull/10
  Work around a bug in geckodriver whereby the pointer location is not
  updated when calling click()
- https://github.com/oleg-andreyev/MinkPhpWebDriver/pull/11
  Trigger a synthetic `change` event rather than pressing TAB after
  sendKeys.
- https://github.com/oleg-andreyev/MinkPhpWebDriver/pull/12
  Work around a bug in geckodriver where an ElementNotInteractable error
  is thrown when attempting to click on a link whose first child is a
  block-level element.

This change moves away from the MoodleSeleium2 Driver and Factory in
favour of the more generic WebDriver and WebDriverFactory.

The changes here also remove the need for the modified syn.js library,
which is no longer used in favour of a true synthetic event.

The customistations in NodeElement are also no longer required as most
of them have been addressed in the php-webdriver/webdriver project, or
are worked around in the oleg-andreyev/MinkPhpWebDriver driver.
