+++ composer.json
@@ -18,15 +18,21 @@
             "role": "Developer"
         }
     ],
+    "repositories": [
+        {
+            "type": "vcs",
+            "url": "https://github.com/andrewnicols/MinkPhpWebDriver.git"
+        }
+    ],
     "require": {
         "php": ">=7.2.0",
         "behat/mink": "~1.8",
         "behat/mink-extension": "~2.3",
         "behat/mink-goutte-driver": "~1.2",
-        "behat/mink-selenium2-driver": "~1.4",
+        "oleg-andreyev/mink-facebook-web-driver": "dev-w3c",
         "symfony/process": "^4.0 || ^5.0",
         "behat/behat": "3.6.*"
-    },
+     },
     "autoload": {
         "psr-0": {
             "Moodle\\BehatExtension": "src/"
