--- src/Moodle/BehatExtension/Driver/Selenium2/README.md
@@ -1,12 +1,0 @@
-Steps to generate the minified file in case we have local modifications and we can't used the minified
-file generated by https://github.com/bitovi/syn. Hopefully a solution for https://github.com/bitovi/syn/issues/30 will
-be merged upstream and we can get rid of these modifications and this README.md
-
-Steps to generate the minified version of syn.js
-================================================
-
-* Checkout the project with our local modifications
-* cd to the directory
-* npm install
-* grunt build
-* The resulting minified file will be stored in dist/syn.min.js
